[Edit] 코드 분리
